fix: update signUp method to add the token during registration
