+++ App/notification/pushNotificationService.js
@@ -1,4 +1,5 @@
 import PushNotification from 'react-native-push-notification';
+import * as firebaseService from '../api/firebaseService';
 
 export const localNotify = (title, description, forumId) => {
   PushNotification.localNotification({
@@ -11,14 +12,15 @@ export const localNotify = (title, description, forumId) => {
   });
 };
 
-export const configurePushNotification = (addUserToken, goToCreatedForum) => {
+export const configurePushNotification = userId => {
   PushNotification.configure({
     onRegister: token => {
-      addUserToken(token.token);
+      firebaseService.addUserToken(token.token, userId);
     },
 
     onNotification: notification => {
-      notification.data.forumId && goToCreatedForum(notification.data.forumId);
+      notification.data.forumId &&
+        firebaseService.goToCreatedForum(notification.data.forumId);
     },
 
     senderID: '1090501687137',
